requirements.txt
Django==1.10.1
Twisted==16.4.1
asgi-redis==0.14.1
asgiref==0.14.0
attrs==16.2.0
autobahn==0.16.0
channels==0.17.2
daphne==0.15.0
dj-database-url==0.4.1
django-ipware==1.1.5
djangorestframework==3.4.6
msgpack-python==0.4.8
ndg-httpsclient==0.4.2
psycopg2==2.6.2
pusher==1.5
pyOpenSSL==16.1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.3
pyuploadcare==2.0.1
redis==2.10.5
requests==2.11.1
service-identity==16.0.0
six==1.10.0
txaio==2.5.1
urllib3==1.17
whitenoise==3.2.1
gunicorn==19.6.0
django-taggit==0.21.2
